requirements.txt
requests~=2.31.0
torch>=1.12.1
numpy>=1.15.4
nltk>=3.5
urllib3>=1.25.8
Unidecode>=1.1.1
ujson~=5.8.0
tqdm~=4.66.1
transformers>=4.2.2
tensorboard
Cython
lmdb>=1.0
boto3~=1.29.5
prettyprint
botocore~=1.32.5
python-dateutil~=2.8.2
setuptools~=68.2.0
python-dotenv~=1.0.0
openai~=1.3.6
together~=0.2.9